Update vkjson for better compatibility
The current vkjson doesn't have a strict checker for discrete enum
values as well as bit masks in VK1.1 in P. So this commit updates the
checker for those special enum targets. e.g. VkFormat. This commit also
add modified the reported extension structure, so that deviceinfo could
consume the extension informaiton.

Bug: b/38269556
Test: adb shell cmd gpu vkjson
Change-Id: I6915dab77474034dd370953d44e5f3386e440775